+++ vite.config.ts
@@ -5,4 +5,9 @@ import react from "@vitejs/plugin-react-swc";
 export default defineConfig({
   plugins: [react()],
   base: "/Advanced-Todo",
+  resolve: {
+    alias: {
+      "@tabler/icons-react": "@tabler/icons-react/dist/esm/icons/index.mjs", // Tabler icons causes slow refresh, so this makes it faster.
+    },
+  },
 });